requirements.txt
streamlit>=1.26.0
streamlit-elements>=0.1.0
streamlit-agraph>=0.0.45
requests>=2.31.0
python-dotenv>=1.0.0
openai>=1.3.0
anthropic>=0.7.0
pydantic>=2.4.0
jsonschema>=4.19.0
pyyaml>=6.0.1